add seed data example

+++ __tests__/main.test.ts
@@ -3,17 +3,32 @@ import request, { Test } from 'supertest';
 import { Server } from 'http';
 import bootServer from '../server';
 import bootDB from '../db';
+import { seedDb, random } from '../__seed__';
+import { UserEntry } from '../models/user';
+import { ProductEntry } from '../models/product';
 
 const port = Number(process.env.TEST_PORT);
 const connectionString = String(process.env.TEST_DB_BASE_URL);
 
 let server: Server;
 let db: Mongoose | undefined;
+let mockUsers: UserEntry[];
+let mockProducts: ProductEntry[];
 
 beforeAll(async () => {
   db = await bootDB(connectionString);
-  await db?.connection.db.dropDatabase();
+  if (db) {
+    await db?.connection.db.dropDatabase();
+    const seedData = await seedDb(db);
+    mockUsers = seedData.users;
+    mockProducts = seedData.products;
+  }
   server = bootServer(port);
+});
+
+test('Mock users and mock products must be present', () => {
+  expect(mockUsers).toHaveLength(50);
+  expect(mockProducts).toHaveLength(200);
 });
 
 describe('POST /register', () => {
@@ -30,10 +45,19 @@ describe('POST /register', () => {
 
   test('returns token', async () => {
     const response = await endpoint.send({
-      firstName: 'Bob', lastName: 'Lovalova', email: 'bob@example.com', password: 'password123',
+      firstName: 'Bob',
+      lastName: 'Lovalova',
+      email: 'bob@example.com',
+      password: 'password123',
     });
     expect(response.status).toBe(201);
     expect(response.body).toHaveProperty('accessToken');
+  });
+
+  test('returns 409 if user exists', async () => {
+    const response = await endpoint.send(mockUsers[random(mockUsers.length)]);
+    expect(response.status).toBe(409);
+    expect(response.body).toHaveProperty('error');
   });
 });
